memory mgmt for EvMsg
When the event is received be:

IRQ handler:
*Reciving* memory buffer -> *Working* memory buffer

Main:
*Working* memory -> call to m_sml.new_message
It will then be moved to the event (EvMsg) data.

EvMsg data ->  *recive action* the recieve action is done
by value which means it will be a copy, the data will then be moved
into the context payload structure,

The idea is that, when a message is completely verified (address, crc) a callback
will be issued, with both header and payload.